[SELC-8445] Moved mapper and fixed tests
