+++ apps/registry-proxy/connector/rest/src/main/java/it/pagopa/selfcare/party/registry_proxy/connector/rest/model/mapper/SearchServiceMapper.java
@@ -1,6 +1,7 @@
-package it.pagopa.selfcare.party.registry_proxy.connector.rest.model.search;
+package it.pagopa.selfcare.party.registry_proxy.connector.rest.model.mapper;
 
 import it.pagopa.selfcare.party.registry_proxy.connector.model.OnboardingIndex;
+import it.pagopa.selfcare.party.registry_proxy.connector.rest.model.search.SearchServiceOnboardingIndex;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
